+++ src/modules/feature/room/RoomMembers/Tables/InvitationTable.vue
