+++ src/modules/feature/room/index.ts
@@ -2,16 +2,16 @@ import RoomGrid from "./RoomGrid.vue";
 import RoomForm from "./RoomForm.vue";
 import RoomMenu from "./RoomMenu.vue";
 import BoardGrid from "./BoardGrid.vue";
-import MembersTable from "./RoomMembers/MembersTable.vue";
-import AddMembers from "./RoomMembers/AddMembers.vue";
-import ChangeRole from "./RoomMembers/ChangeRole.vue";
+import MembersTable from "./RoomMembers/Tables/MembersTable.vue";
+import AddMembers from "./RoomMembers/Dialogs/AddMembers.vue";
+import ChangeRole from "./RoomMembers/Dialogs/ChangeRole.vue";
 import RoomCopyInfoDialog from "./RoomCopyInfoDialog.vue";
-import Members from "./RoomMembers/Members.vue";
-import Invitations from "./RoomMembers/Invitations.vue";
-import Confirmations from "./RoomMembers/Confirmations.vue";
-import InviteMembersDialog from "./RoomMembers/InviteMembersDialog.vue";
-import InvitationTable from "./RoomMembers/InvitationTable.vue";
-import ConfirmationTable from "./RoomMembers/ConfirmationTable.vue";
+import Members from "./RoomMembers/Tabs/Members.vue";
+import Invitations from "./RoomMembers/Tabs/Invitations.vue";
+import Confirmations from "./RoomMembers/Tabs/Confirmations.vue";
+import InviteMembersDialog from "./RoomMembers/Dialogs/InviteMembersDialog.vue";
+import InvitationTable from "./RoomMembers/Tables/InvitationTable.vue";
+import ConfirmationTable from "./RoomMembers/Tables/ConfirmationTable.vue";
 
 export {
 	AddMembers,
